stash introduce ItemMod
This allows us to handle item mods as well as properties. Slow as hell
though, I see the check command time nearly double. Not an issue, its
always easy to switch to a faster regexp engine. Plus I'm probably doing
a shit ton of allocations that don't need to happen...