db IndexQuery LookupItemsMultiModStrideLength to 1
This allows us to generate an ItemStoreQuery from an IndexQuery which
has equivalent semantics without making any changes to the way
ItemStoreQuery operated apart from that it performs an intersection
every stride.

Hopefully, this is temporary and will be reverted once we have the tests
as a solid baseline for changes.