dbTest introduce 140 update testSet + benchmarks fast and slow
This allows for two classes of benchmarks.
The 'Fast' benchmarks provide relatively quick feedback on if an
improvement was worthwhile and look at 11 Responses.
The 'Slow' benchmarks provide an exhaustive test of whether or not a
change is effective under significantly higher load. 140 responses means
it takes quite awhile to run.